Module nvapi_hi::sys

Modules

dispcontrol

driverapi
The display driver APIs are used to retrieve information about the display driver.

gpu
The GPU APIs retrieve and control various attributes of the GPU, such as outputs, VBIOS revision, APG rate, frame buffer size, and thermal settings.

handles
NVAPI Handles - These handles are retrieved from various calls and passed in to others in NvAPI These are meant to be opaque types. Do not assume they correspond to indices, HDCs, display indexes or anything else.

nvapi

nvid

status

types

Structs

ArgumentRangeError
Error type indicating a raw value is out of the range of known enum values.

Array
Hacks a Debug impl onto large arrays.

BoolU32
A boolean containing reserved bits

NV_RECT

NvGUID

NvSBox

Enums

Api

Status
NvAPI Status Values

Constants

NVAPI_ADVANCED_DISPLAY_HEADS

NVAPI_GENERIC_STRING_MAX

NVAPI_LONG_STRING_MAX

NVAPI_MAX_ACPI_IDS

NVAPI_MAX_AUDIO_DEVICES

NVAPI_MAX_AVAILABLE_GPU_TOPOLOGIES

NVAPI_MAX_AVAILABLE_SLI_GROUPS

NVAPI_MAX_DISPLAYS

NVAPI_MAX_DISPLAY_HEADS

NVAPI_MAX_GPU_PER_TOPOLOGY

NVAPI_MAX_GPU_TOPOLOGIES

NVAPI_MAX_HEADS_PER_GPU

NVAPI_MAX_LOGICAL_GPUS

NVAPI_MAX_PHYSICAL_BRIDGES

NVAPI_MAX_PHYSICAL_GPUS

NVAPI_MAX_VIEW_MODES

NVAPI_PHYSICAL_GPUS

NVAPI_SHORT_STRING_MAX

NVAPI_SYSTEM_HWBC_INVALID_ID

NVAPI_SYSTEM_MAX_DISPLAYS

NVAPI_SYSTEM_MAX_HWBCS

NV_FALSE

NV_MAX_HEADS
Maximum heads, each with NVAPI_DESKTOP_RES resolution

NV_MAX_VID_PROFILES
Maximum number of output video profiles supported

NV_MAX_VID_STREAMS
Maximum number of input video streams, each with a NVAPI_VIDEO_SRC_INFO

NV_TRUE

Functions

GET_NVAPI_SIZE

GET_NVAPI_VERSION

MAKE_NVAPI_VERSION
NvAPI Version Definition

long_string

nvapi_QueryInterface

short_string

status_result
Treat NVAPI_OK as Ok(()) and all else as an Err(..).

string

Type Definitions

NvAPI_LongString

NvAPI_ShortString

NvAPI_Status
NvAPI Status Values

NvAPI_String

NvBool

NvLUID

Result
The result of a fallible NVAPI call.