Module nvapi_hi::sys::dispcontrol

Functions

NvAPI_DISP_GetAssociatedUnAttachedNvidiaDisplayHandle⚠
This function returns the handle of an unattached NVIDIA display that is associated with the given display name (such as "\DISPLAY1").

NvAPI_EnumNvidiaDisplayHandle⚠
This function returns the handle of the NVIDIA display specified by the enum index (thisEnum). The client should keep enumerating until it returns NVAPI_END_ENUMERATION.

NvAPI_EnumNvidiaUnAttachedDisplayHandle⚠
This function returns the handle of the NVIDIA unattached display specified by the enum index (thisEnum). The client should keep enumerating until it returns error.

NvAPI_GetAssociatedNvidiaDisplayHandle⚠
This function returns the handle of the NVIDIA display that is associated with the given display "name" (such as "\.\DISPLAY1").

Type Definitions

DISP_GetAssociatedUnAttachedNvidiaDisplayHandleFn

EnumNvidiaDisplayHandleFn

EnumNvidiaUnAttachedDisplayHandleFn

GetAssociatedNvidiaDisplayHandleFn