Module nvapi_hi::sys::driverapi

The display driver APIs are used to retrieve information about the display driver.

Modules

private
Undocumented API

Structs

NV_DISPLAY_DRIVER_MEMORY_INFO_V1
Used in NvAPI_GPU_GetMemoryInfo().

NV_DISPLAY_DRIVER_MEMORY_INFO_V2
Used in NvAPI_GPU_GetMemoryInfo().

NV_DISPLAY_DRIVER_MEMORY_INFO_V3
Used in NvAPI_GPU_GetMemoryInfo().

Constants

NV_DISPLAY_DRIVER_MEMORY_INFO_VER

NV_DISPLAY_DRIVER_MEMORY_INFO_VER_1

NV_DISPLAY_DRIVER_MEMORY_INFO_VER_2

NV_DISPLAY_DRIVER_MEMORY_INFO_VER_3

Functions

NvAPI_GPU_GetMemoryInfo⚠
This function retrieves the available driver memory footprint for the specified GPU. If the GPU is in TCC Mode, only dedicatedVideoMemory will be returned in pMemoryInfo (NV_DISPLAY_DRIVER_MEMORY_INFO).

NvAPI_SYS_GetDriverAndBranchVersion⚠
This API returns display driver version and driver-branch string.

Type Definitions

GPU_GetMemoryInfoFn

NV_DISPLAY_DRIVER_MEMORY_INFO

SYS_GetDriverAndBranchVersionFn