Module nvapi_hi::sys::gpu

The GPU APIs retrieve and control various attributes of the GPU, such as outputs, VBIOS revision, APG rate, frame buffer size, and thermal settings.

Modules

clock
The GPU clock control APIs are used to get and set individual clock domains on a per-GPU basis.

cooler
The GPU cooler APIs are used to get and set the fan level or equivalent cooler levels for various target devices associated with the GPU.

display

power

private
Undocumented API

pstate
The GPU performance state APIs are used to get and set various performance levels on a per-GPU basis. P-States are GPU active/executing performance capability and power consumption states.

thermal
The GPU thermal control APIs are used to get temperature levels from the various thermal sensors associated with the GPU.

Enums

SystemType
Used in NvAPI_GPU_GetSystemType()

Constants

NV_SYSTEM_TYPE_DESKTOP

NV_SYSTEM_TYPE_LAPTOP

NV_SYSTEM_TYPE_UNKNOWN

Functions

NvAPI_EnumPhysicalGPUs⚠
This function returns an array of physical GPU handles. Each handle represents a physical GPU present in the system. That GPU may be part of an SLI configuration, or may not be visible to the OS directly.

NvAPI_GPU_GetFullName⚠
This function retrieves the full GPU name as an ASCII string - for example, "Quadro FX 1400".

NvAPI_GPU_GetGpuCoreCount⚠
Retrieves the total number of cores defined for a GPU. Returns 0 on architectures that don't define GPU cores.

NvAPI_GPU_GetPCIIdentifiers⚠
This function returns the PCI identifiers associated with this GPU.

NvAPI_GPU_GetPhysicalFrameBufferSize⚠
This function returns the physical size of framebuffer in KB. This does NOT include any system RAM that may be dedicated for use by the GPU.

NvAPI_GPU_GetShaderSubPipeCount⚠
This function retrieves the number of Shader SubPipes on the GPU On newer architectures, this corresponds to the number of SM units

NvAPI_GPU_GetSystemType⚠
This function identifies whether the GPU is a notebook GPU or a desktop GPU.

NvAPI_GPU_GetVbiosVersionString⚠
This function returns the full video BIOS version string in the form of xx.xx.xx.xx.yy where - xx numbers come from NvAPI_GPU_GetVbiosRevision() and - yy comes from NvAPI_GPU_GetVbiosOEMRevision().

Type Definitions

EnumPhysicalGPUsFn

GPU_GetFullNameFn

GPU_GetGpuCoreCountFn

GPU_GetPCIIdentifiersFn

GPU_GetPhysicalFrameBufferSizeFn

GPU_GetShaderSubPipeCountFn

GPU_GetSystemTypeFn

GPU_GetVbiosVersionStringFn

NV_SYSTEM_TYPE
Used in NvAPI_GPU_GetSystemType()