Module nvapi_hi::sys::gpu::cooler::private

Undocumented API

Structs

NV_GPU_COOLER_POLICY_LEVEL

NV_GPU_COOLER_POLICY_TABLE_V1

NV_GPU_COOLER_SETTINGS_COOLER

NV_GPU_COOLER_SETTINGS_V1

NV_GPU_SETCOOLER_LEVEL_COOLER

NV_GPU_SETCOOLER_LEVEL_V1

Enums

CoolerActivityLevel

CoolerControl

CoolerController

CoolerPolicy

CoolerTarget

CoolerType

Constants

NVAPI_ACTIVE

NVAPI_COOLER_CONTROLLER_ADI

NVAPI_COOLER_CONTROLLER_INTERNAL

NVAPI_COOLER_CONTROLLER_NONE

NVAPI_COOLER_CONTROL_NONE

NVAPI_COOLER_CONTROL_TOGGLE
ON/OFF

NVAPI_COOLER_CONTROL_VARIABLE
Suppports variable control.

NVAPI_COOLER_POLICY_HYBRID
Hybrid of performance and temperature levels.

NVAPI_COOLER_POLICY_MANUAL
Manual adjustment of cooler level. Gets applied right away independent of temperature or performance level.

NVAPI_COOLER_POLICY_NONE

NVAPI_COOLER_POLICY_PERF
GPU performance controls the cooler level.

NVAPI_COOLER_POLICY_SILENT
Fan turns off at idle, default of MSI Gaming X

NVAPI_COOLER_POLICY_TEMPERATURE_CONTINUOUS
Cooler level adjusted at continuous thermal levels.

NVAPI_COOLER_POLICY_TEMPERATURE_DISCRETE
Discrete thermal levels control the cooler level.

NVAPI_COOLER_POLICY_UNKNOWN_32
Apparently a default of some GPUs

NVAPI_COOLER_TARGET_ALL
This cooler cools all of the components related to its target gpu.

NVAPI_COOLER_TARGET_GPU

NVAPI_COOLER_TARGET_MEMORY

NVAPI_COOLER_TARGET_NONE

NVAPI_COOLER_TARGET_POWER_SUPPLY

NVAPI_COOLER_TYPE_FAN

NVAPI_COOLER_TYPE_LIQUID_NO2

NVAPI_COOLER_TYPE_NONE

NVAPI_COOLER_TYPE_WATER

NVAPI_INACTIVE

NVAPI_MAX_COOLERS_PER_GPU

NVAPI_MAX_COOLER_LEVEL

NVAPI_MAX_COOLER_LEVELS

NVAPI_MIN_COOLER_LEVEL

NV_GPU_COOLER_POLICY_TABLE_VER

NV_GPU_COOLER_POLICY_TABLE_VER_1

NV_GPU_COOLER_SETTINGS_VER

NV_GPU_COOLER_SETTINGS_VER_1

NV_GPU_SETCOOLER_LEVEL_VER

NV_GPU_SETCOOLER_LEVEL_VER_1

Functions

NvAPI_GPU_GetCoolerPolicyTable⚠
Undocumented function. Retrieves the table of cooler and policy levels for the selected policy. Supported only for NVAPI_COOLER_POLICY_PERF.

NvAPI_GPU_GetCoolerSettings⚠
Undocumented function. Retrieves the cooler information of all coolers or a specific cooler associated with the selected GPU.

NvAPI_GPU_RestoreCoolerPolicyTable⚠
Undocumented function. Restores the perf table policy levels to the defaults.

NvAPI_GPU_RestoreCoolerSettings⚠
Undocumented function. Restore the modified cooler settings to NVIDIA defaults.

NvAPI_GPU_SetCoolerLevels⚠
Undocumented function. Set the cooler levels for all coolers or a specific cooler associated with the selected GPU.

NvAPI_GPU_SetCoolerPolicyTable⚠
Undocumented function. Restore the modified cooler settings to NVIDIA defaults. Supported only for NVAPI_COOLER_POLICY_PERF.

Type Definitions

GPU_GetCoolerPolicyTableFn

GPU_GetCoolerSettingsFn

GPU_RestoreCoolerPolicyTableFn

GPU_RestoreCoolerSettingsFn

GPU_SetCoolerLevelsFn

GPU_SetCoolerPolicyTableFn

NV_COOLER_ACTIVITY_LEVEL

NV_COOLER_CONTROL

NV_COOLER_CONTROLLER

NV_COOLER_POLICY

NV_COOLER_TARGET

NV_COOLER_TYPE

NV_GPU_COOLER_POLICY_TABLE

NV_GPU_COOLER_SETTINGS

NV_GPU_SETCOOLER_LEVEL