Struct nvapi_hi::sys::gpu::cooler::private::NV_GPU_COOLER_POLICY_LEVEL

#[repr(C)]

pub struct NV_GPU_COOLER_POLICY_LEVEL {
    pub levelId: u32,
    pub currentLevel: u32,
    pub defaultLevel: u32,
}

Fields
levelId: u32

level indicator for a policy

currentLevel: u32

new cooler level for the selected policy level indicator.

defaultLevel: u32

default cooler level for the selected policy level indicator.

Methods

impl NV_GPU_COOLER_POLICY_LEVEL

pub fn zeroed() -> NV_GPU_COOLER_POLICY_LEVEL