Module nvapi_hi::sys::gpu::display

Structs

ConnectedIdsFlags
Argument to NvAPI_GPU_GetConnectedDisplayIds

DisplayIdsFlags
Bitfield in NV_GPU_DISPLAYIDS

NV_GPU_DISPLAYIDS

Enums

MonitorConnectorType

Constants

NV_GPU_CONNECTED_IDS_FLAG_EXCLUDE_MST
Excludes devices that are part of the multi stream topology.

NV_GPU_CONNECTED_IDS_FLAG_FAKE
Get devices that includes the fake connected monitors

NV_GPU_CONNECTED_IDS_FLAG_LIDSTATE
Get devices such that to reflect the Lid State

NV_GPU_CONNECTED_IDS_FLAG_SLI
Get devices such that those can be selected in an SLI configuration

NV_GPU_CONNECTED_IDS_FLAG_UNCACHED
Get uncached connected devices

NV_GPU_DISPLAYIDS_FLAGS_ACTIVE
This display is being actively driven

NV_GPU_DISPLAYIDS_FLAGS_CLUSTER
This display is the representative display

NV_GPU_DISPLAYIDS_FLAGS_CONNECTED
This display is connected

NV_GPU_DISPLAYIDS_FLAGS_DYNAMIC
This display is part of MST topology and it's a dynamic

NV_GPU_DISPLAYIDS_FLAGS_MST_ROOT_NODE
This displayID belongs to a multi stream enabled connector(root node). Note that when multi stream is enabled and a single multi stream capable monitor is connected to it, the monitor will share the display id with the RootNode.

NV_GPU_DISPLAYIDS_FLAGS_OS_VISIBLE
This display is reported to the OS

NV_GPU_DISPLAYIDS_FLAGS_PHYSICALLY_CONNECTED
this display is a phycially connected display; Valid only when is Connected bit is set

NV_GPU_DISPLAYIDS_FLAGS_RESERVED
must be zero

NV_GPU_DISPLAYIDS_FLAGS_RESERVED_INTERNAL
Do not use

NV_GPU_DISPLAYIDS_FLAGS_WIRELESS
This display is wireless

NV_GPU_DISPLAYIDS_VER

NV_GPU_DISPLAYIDS_VER1

NV_GPU_DISPLAYIDS_VER2

NV_MONITOR_CONN_TYPE_COMPONENT

NV_MONITOR_CONN_TYPE_COMPOSITE

NV_MONITOR_CONN_TYPE_DP

NV_MONITOR_CONN_TYPE_DVI

NV_MONITOR_CONN_TYPE_HDMI

NV_MONITOR_CONN_TYPE_LVDS

NV_MONITOR_CONN_TYPE_SVIDEO

NV_MONITOR_CONN_TYPE_UNINITIALIZED

NV_MONITOR_CONN_TYPE_UNKNOWN

NV_MONITOR_CONN_TYPE_VGA

Functions

NvAPI_GPU_GetAllDisplayIds⚠
This API returns display IDs for all possible outputs on the GPU. For DPMST connector, it will return display IDs for all the video sinks in the topology.

NvAPI_GPU_GetConnectedDisplayIds⚠
Due to space limitation NvAPI_GPU_GetConnectedOutputs can return maximum 32 devices, but this is no longer true for DPMST. NvAPI_GPU_GetConnectedDisplayIds will return all the connected display devices in the form of displayIds for the associated hPhysicalGpu. This function can accept set of flags to request cached, uncached, sli and lid to get the connected devices. Default value for flags will be cached.

Type Definitions

GPU_GetAllDisplayIds

GPU_GetConnectedDisplayIds

NV_GPU_CONNECTED_IDS_FLAG
Argument to NvAPI_GPU_GetConnectedDisplayIds

NV_GPU_DISPLAYIDS_FLAGS
Bitfield in NV_GPU_DISPLAYIDS

NV_MONITOR_CONN_TYPE