Struct nvapi_hi::sys::gpu::display::ConnectedIdsFlags

pub struct ConnectedIdsFlags { /* fields omitted */ }

Argument to NvAPI_GPU_GetConnectedDisplayIds

Methods

impl ConnectedIdsFlags

pub const UNCACHED: ConnectedIdsFlags

UNCACHED: ConnectedIdsFlags = ConnectedIdsFlags{bits: 1,}

Get uncached connected devices

pub const SLI: ConnectedIdsFlags

SLI: ConnectedIdsFlags = ConnectedIdsFlags{bits: 2,}

Get devices such that those can be selected in an SLI configuration

pub const LID_STATE: ConnectedIdsFlags

LID_STATE: ConnectedIdsFlags = ConnectedIdsFlags{bits: 4,}

Get devices such that to reflect the Lid State

pub const FAKE: ConnectedIdsFlags

FAKE: ConnectedIdsFlags = ConnectedIdsFlags{bits: 8,}

Get devices that includes the fake connected monitors

pub const EXCLUDE_MST: ConnectedIdsFlags

EXCLUDE_MST: ConnectedIdsFlags = ConnectedIdsFlags{bits: 16,}

Excludes devices that are part of the multi stream topology.

Important traits for ConnectedIdsFlags
impl Iterator for ConnectedIdsFlags type Item = ConnectedIdsFlags;

pub fn empty() -> ConnectedIdsFlags

Returns an empty set of flags.

Important traits for ConnectedIdsFlags
impl Iterator for ConnectedIdsFlags type Item = ConnectedIdsFlags;

pub fn all() -> ConnectedIdsFlags

Returns the set containing all flags.

pub fn bits(&self) -> u32

Returns the raw value of the flags currently stored.

pub fn from_bits(bits: u32) -> Option<ConnectedIdsFlags>

Convert from underlying bit representation, unless that representation contains bits that do not correspond to a flag.

Important traits for ConnectedIdsFlags
impl Iterator for ConnectedIdsFlags type Item = ConnectedIdsFlags;

pub fn from_bits_truncate(bits: u32) -> ConnectedIdsFlags

Convert from underlying bit representation, dropping any bits that do not correspond to flags.

pub fn is_empty(&self) -> bool

Returns true if no flags are currently stored.

pub fn is_all(&self) -> bool

Returns true if all flags are currently set.

pub fn intersects(&self, other: ConnectedIdsFlags) -> bool

Returns true if there are flags common to both self and other.

pub fn contains(&self, other: ConnectedIdsFlags) -> bool

Returns true all of the flags in other are contained within self.

pub fn insert(&mut self, other: ConnectedIdsFlags)

Inserts the specified flags in-place.

pub fn remove(&mut self, other: ConnectedIdsFlags)

Removes the specified flags in-place.

pub fn toggle(&mut self, other: ConnectedIdsFlags)

Toggles the specified flags in-place.

pub fn set(&mut self, other: ConnectedIdsFlags, value: bool)

Inserts or removes the specified flags depending on the passed value.