Function nvapi_hi::sys::gpu::NvAPI_EnumPhysicalGPUs

pub unsafe fn NvAPI_EnumPhysicalGPUs(
    nvGPUHandle: *mut [NvPhysicalGpuHandle; 64], 
    pGpuCount: *mut u32
) -> i32

This function returns an array of physical GPU handles. Each handle represents a physical GPU present in the system. That GPU may be part of an SLI configuration, or may not be visible to the OS directly.

At least one GPU must be present in the system and running an NVIDIA display driver.

The array nvGPUHandle will be filled with physical GPU handle values. The returned gpuCount determines how many entries in the array are valid.

Note: In drivers older than 105.00, all physical GPU handles get invalidated on a modeset. So the calling applications need to renum the handles after every modeset. With drivers 105.00 and up, all physical GPU handles are constant. Physical GPU handles are constant as long as the GPUs are not physically moved and the SBIOS VGA order is unchanged.

For GPU handles in TCC MODE please use NvAPI_EnumTCCPhysicalGPUs()