Module nvapi_hi::sys::i2c

I2C API - Provides ability to read or write data using I2C protocol. These APIs allow I2C access only to DDC monitors

Modules

private
Undocumented API

Structs

NV_I2C_INFO_V1
Used in NvAPI_I2CRead() and NvAPI_I2CWrite()

NV_I2C_INFO_V2
Used in NvAPI_I2CRead() and NvAPI_I2CWrite()

NV_I2C_INFO_V3
Used in NvAPI_I2CRead() and NvAPI_I2CWrite()

Enums

I2cSpeed

Constants

NVAPI_DISPLAY_DEVICE_MASK_MAX

NVAPI_I2C_SPEED_100KHZ

NVAPI_I2C_SPEED_10KHZ

NVAPI_I2C_SPEED_200KHZ

NVAPI_I2C_SPEED_33KHZ

NVAPI_I2C_SPEED_3KHZ

NVAPI_I2C_SPEED_400KHZ

NVAPI_I2C_SPEED_DEFAULT

NVAPI_I2C_SPEED_DEPRECATED

NVAPI_MAX_SIZEOF_I2C_DATA_BUFFER

NVAPI_MAX_SIZEOF_I2C_REG_ADDRESS

NV_I2C_INFO_VER

NV_I2C_INFO_VER1

NV_I2C_INFO_VER2

NV_I2C_INFO_VER3

Functions

NvAPI_I2CRead⚠
This function reads the data buffer from the I2C port. The I2C request must be for a DDC port: pI2cInfo->bIsDDCPort = 1.

NvAPI_I2CWrite⚠
This function writes the data buffer to the I2C port.

Type Definitions

NV_I2C_INFO

NV_I2C_SPEED

NvAPI_I2CReadFn

NvAPI_I2CWriteFn