Function nvapi_hi::sys::i2c::NvAPI_I2CRead

pub unsafe fn NvAPI_I2CRead(
    hPhysicalGpu: NvPhysicalGpuHandle, 
    pI2cInfo: *mut NV_I2C_INFO_V3
) -> i32

This function reads the data buffer from the I2C port. The I2C request must be for a DDC port: pI2cInfo->bIsDDCPort = 1.

A data buffer size larger than 16 bytes may be rejected if a register address is specified. In such a case, NVAPI_ARGUMENT_EXCEED_MAX_SIZE would be returned.

If a register address is specified (i.e. regAddrSize is positive), then the transaction will be performed in the combined format described in the I2C specification. The register address will be written, followed by reading into the data buffer.

Returns

NVAPI_OK: Completed request

NVAPI_ERROR: Miscellaneous error occurred.

NVAPI_HANDLE_INVALIDATED: Handle passed has been invalidated (see user guide).

NVAPI_EXPECTED_PHYSICAL_GPU_HANDLE: Handle passed is not a physical GPU handle.

NVAPI_INCOMPATIBLE_STRUCT_VERSION: Structure version is not supported.

NVAPI_INVALID_ARGUMENT: argument does not meet specified requirements

NVAPI_ARGUMENT_EXCEED_MAX_SIZE: an argument exceeds the maximum