Module nvapi_hi::sys::status

Enums

Status
NvAPI Status Values

Constants

NVAPI_ACCESS_DENIED
No access to the caller.

NVAPI_ADVANCED_DISPLAY_TOPOLOGY_REQUIRED
Information can't be returned due to "advanced display topology".

NVAPI_API_IN_USE
Atleast an API is still being called

NVAPI_API_NOT_INITIALIZED
NvAPI_Initialize has not been called (successfully)

NVAPI_ARGUMENT_EXCEED_MAX_SIZE
Argument exceeds the expected size.

NVAPI_CLUSTER_ALREADY_EXISTS
A cluster is already defined with the given configuration.

NVAPI_D3D10_1_LIBRARY_NOT_FOUND
d3d10_1.dll cannot be loaded.

NVAPI_D3D11_LIBRARY_NOT_FOUND
d3d11.dll/d3d11_beta.dll cannot be loaded.

NVAPI_D3D_DEVICE_LOST
D3D device status is D3DERR_DEVICELOST or D3DERR_DEVICENOTRESET - the user has to reset the device.

NVAPI_D3D_DEVICE_NOT_REGISTERED
The user tried to use a deferred context without registering the device first

NVAPI_DATATYPE_MISMATCH
Data Type mismatch

NVAPI_DATA_NOT_FOUND
Requested data was not found

NVAPI_DEFAULT_STEREO_PROFILE_DOES_NOT_EXIST
Default stereo profile does not exist

NVAPI_DEFAULT_STEREO_PROFILE_IS_NOT_DEFINED
Default stereo profile is not currently defined

NVAPI_DEVICE_BUSY
Device is busy; request not fulfilled

NVAPI_DEVICE_SWITCHING_NOT_ALLOWED
Inhibit is ON due to one of the flags in NV_GPU_DISPLAY_CHANGE_INHIBIT or SLI active.

NVAPI_DISPLAYCONFIG_VALIDATION_FAILED
Display topology is not valid so the driver cannot do a mode set on this configuration.

NVAPI_DISPLAY_IN_USE
DirectMode Display is already in use

NVAPI_DPMST_CHANGED
Display Port Multi-Stream topology has been changed.

NVAPI_DPMST_DISPLAY_ID_EXPECTED
The input display id is not that of a multi stream enabled connector or a display device in a multi stream topology

NVAPI_DRIVER_NOTRUNNING
Driver is not running.

NVAPI_DRIVER_RUNNING
Driver is running.

NVAPI_ECID_KEY_VERIFICATION_FAILED
The encrypted public key verification has failed.

NVAPI_ECID_SIGN_ALGO_UNSUPPORTED
The specified signing algorithm is not supported. Either an incorrect value was entered or the current installed driver/hardware does not support the input value.

NVAPI_END_ENUMERATION
No more items to enumerate

NVAPI_ERROR
Generic error

NVAPI_ERROR_DRIVER_RELOAD_REQUIRED
A driver reload is required to apply these settings.

NVAPI_EXECUTABLE_ALREADY_IN_USE
Application already exists in the other profile.

NVAPI_EXECUTABLE_NOT_FOUND
Application not found in the Profile.

NVAPI_EXECUTABLE_PATH_IS_AMBIGUOUS
The path provided was too short to determine the correct NVDRS_APPLICATION

NVAPI_EXPECTED_ANALOG_DISPLAY
Expected an analog display

NVAPI_EXPECTED_COMPUTE_GPU_HANDLE
Setting Physx GPU requires that the GPU is compute-capable.

NVAPI_EXPECTED_DIGITAL_FLAT_PANEL
Expected a digital flat panel.

NVAPI_EXPECTED_DISPLAY_HANDLE
Expected an NV display handle for one or more parameters

NVAPI_EXPECTED_LOGICAL_GPU_HANDLE
Expected a logical GPU handle for one or more parameters

NVAPI_EXPECTED_NON_PRIMARY_DISPLAY_HANDLE
The handle corresponds to GDIPrimary.

NVAPI_EXPECTED_PHYSICAL_GPU_HANDLE
Expected a physical GPU handle for one or more parameters

NVAPI_EXPECTED_TV_DISPLAY
Expected a TV output display

NVAPI_EXPECTED_TV_DISPLAY_ON_DCONNECTOR
Expected a TV output on the D Connector - HDTV_EIAJ4120.

NVAPI_EXPECTED_UNATTACHED_DISPLAY_HANDLE
Expected an unattached display handle as one of the input parameters.

NVAPI_FILE_NOT_FOUND
The file was not found.

NVAPI_FIRMWARE_OUT_OF_DATE
The device's firmware is out of date.

NVAPI_FIRMWARE_REVISION_NOT_SUPPORTED
The device's firmware is not supported.

NVAPI_FUNCTION_NOT_FOUND
Couldn't find the function in the loaded DLL.

NVAPI_GPU_IN_DEBUG_MODE
GPU is in debug mode, OC is NOT allowed.

NVAPI_GPU_WORKSTATION_FEATURE_INCOMPLETE
The requested workstation feature set has incomplete driver internal allocation resources

NVAPI_HANDLE_INVALIDATED
The handle is no longer valid (likely due to GPU or display re-configuration)

NVAPI_HDCP_DISABLED
When a non-HDCP capable HMD is connected, we would inform user by this code.

NVAPI_HDCP_ENCRYPTION_FAILED
HDCP Encryption Failed for the device. Would be applicable when the device is HDCP Capable.

NVAPI_ID_OUT_OF_RANGE
The DisplayId corresponds to a display which is not within the normal outputId range.

NVAPI_IMPLICIT_SET_GPU_TOPOLOGY_CHANGE_NOT_ALLOWED
There is no implicit GPU topology active. Use NVAPI_SetHybridMode to change topology.

NVAPI_INCOMPATIBLE_AUDIO_DRIVER
Older audio driver version than required

NVAPI_INCOMPATIBLE_STRUCT_VERSION
An argument's structure version is not supported

NVAPI_INCORRECT_VENDOR
NVAPI cannot be called by this vendor

NVAPI_INSTRUMENTATION_DISABLED
OpenGL Expert is supported, but driver instrumentation is currently disabled

NVAPI_INSUFFICIENT_BUFFER
Input buffer is insufficient to hold the contents.

NVAPI_INVALID_ADDRESS
Address is outside of valid range.

NVAPI_INVALID_ARGUMENT
The argument/parameter value is not valid or NULL.

NVAPI_INVALID_CALL
The method call is invalid. For example, a method's parameter may not be a valid pointer.

NVAPI_INVALID_COMBINATION
The combination of parameters is not valid.

NVAPI_INVALID_CONFIGURATION
The requested action cannot be performed in the current state.

NVAPI_INVALID_DIRECT_MODE_DISPLAY
DirectMode not Enabled on the Display

NVAPI_INVALID_DISPLAY_ID
The input display id is not valid or the monitor associated to it does not support the current operation

NVAPI_INVALID_HANDLE
Invalid handle

NVAPI_INVALID_HYBRID_MODE
The function is not supported with the current Hybrid mode.

NVAPI_INVALID_PERF_LEVEL
Invalid perf level

NVAPI_INVALID_POINTER
An invalid pointer, usually NULL, was passed as a parameter

NVAPI_INVALID_SYNC_TOPOLOGY
Invalid displays passed in the NV_GSYNC_DISPLAY pointer.

NVAPI_INVALID_USER_PRIVILEGE
Current User is not Admin.

NVAPI_LIBRARY_NOT_FOUND
NVAPI support library cannot be loaded.

NVAPI_LICENSE_CALLER_AUTHENTICATION_FAILED
The caller is not authorized to modify the License.

NVAPI_MATCHING_DEVICE_NOT_FOUND
The input does not match any of the available devices.

NVAPI_MAX_DISPLAY_LIMIT_REACHED
GPU's Max Display Limit has Reached

NVAPI_MIXED_TARGET_TYPES
The target types are not all the same

NVAPI_MODE_CHANGE_FAILED
An attempt to do a display resolution mode change has failed.

NVAPI_MOSAIC_NOT_ACTIVE
The requested action cannot be performed without Mosaic being enabled.

NVAPI_NOT_SUPPORTED
Requested feature is not supported in the selected GPU

NVAPI_NO_ACTIVE_SLI_TOPOLOGY
SLI is not active on this device.

NVAPI_NO_CONNECTOR_FOUND
No connector on GPU found.

NVAPI_NO_GL_EXPERT
OpenGL Expert is not supported by the current drivers

NVAPI_NO_GL_NSIGHT
OpenGL does not support Nsight

NVAPI_NO_IMPLEMENTATION
not implemented in current driver installation

NVAPI_NO_VIDLINK
No SLI video bridge is present

NVAPI_NVIDIA_DEVICE_NOT_FOUND
No NVIDIA display driver, or NVIDIA GPU driving a display, was found.

NVAPI_NVIDIA_DISPLAY_NOT_FOUND
No display found on Nvidia GPU(s).

NVAPI_NV_PERSIST_FILE_NOT_FOUND
NV persist file is not found

NVAPI_OK
Success. Request is completed.

NVAPI_OPENGL_CONTEXT_NOT_CURRENT
No NVIDIA OpenGL context is current (but needs to be)

NVAPI_OUT_OF_MEMORY
Could not allocate sufficient memory to complete the call.

NVAPI_PCLK_LIMITATION_FAILED
Provided mode is over sink device pclk limitation.

NVAPI_PERSIST_DATA_NOT_FOUND
NV persist data is not found

NVAPI_PORTID_NOT_FOUND
No port ID was found for the I2C transaction

NVAPI_PRIV_SEC_VIOLATION
Priv security violation, improper access to a secured register.

NVAPI_PROFILE_NAME_EMPTY
Profile name is empty.

NVAPI_PROFILE_NAME_IN_USE
Profile name is duplicated.

NVAPI_PROFILE_NOT_FOUND
Profile is not found.

NVAPI_PROFILE_REMOVED
The profile passed as parameter has been removed and is no longer valid.

NVAPI_REQUEST_USER_TO_CLOSE_NON_MIGRATABLE_APPS
Prompt the user to close all non-migratable applications.

NVAPI_REQUEST_USER_TO_DISABLE_DWM
The user should disable DWM before calling NvAPI.

NVAPI_REQUIRES_REBOOT
NVAPI requires a reboot for the settings to take effect

NVAPI_RESOURCE_NOT_ACQUIRED
Head or SourceId was not reserved for the VR Display before doing the Modeset.

NVAPI_SETTING_NOT_FOUND
Setting is not found.

NVAPI_SETTING_SIZE_TOO_LARGE
Setting size is too large.

NVAPI_SET_NOT_ALLOWED
Intended setting is not allowed.

NVAPI_SHARE_RESOURCE_RELOCATED
The surface is relocated away from video memory.

NVAPI_SLI_RENDERING_MODE_NOTALLOWED
Setup of SLI rendering mode is not possible right now.

NVAPI_STEREO_FRUSTUM_ADJUST_MODE_NOT_SUPPORTED
The given frustum adjust mode is not supported.

NVAPI_STEREO_HANDSHAKE_NOT_DONE
Call failed as stereo handshake not completed.

NVAPI_STEREO_INIT_ACTIVATION_NOT_DONE
Call failed because InitActivation was not called.

NVAPI_STEREO_INVALID_DEVICE_INTERFACE
Invalid device interface.

NVAPI_STEREO_NOT_ENABLED
Stereo is not enabled and the function needed it to execute completely.

NVAPI_STEREO_NOT_INITIALIZED
The Stereo part of NVAPI failed to initialize completely. Check if the stereo driver is installed.

NVAPI_STEREO_NOT_TURNED_ON
Stereo is not turned on and the function needed it to execute completely.

NVAPI_STEREO_PARAMETER_OUT_OF_RANGE
Separation percentage or JPEG image capture quality is out of [0-100] range.

NVAPI_STEREO_REGISTRY_ACCESS_FAILED
Access to stereo-related registry keys or values has failed.

NVAPI_STEREO_REGISTRY_PROFILE_TYPE_NOT_SUPPORTED
The given registry profile type is not supported.

NVAPI_STEREO_REGISTRY_VALUE_NOT_SUPPORTED
The given registry value is not supported.

NVAPI_STREAM_IS_OUT_OF_SYNC
While playing secure audio stream, stream goes out of sync

NVAPI_STRING_TOO_SMALL
The pre-allocated string is too small to hold the result.

NVAPI_SYNC_MASTER_NOT_FOUND
The requested action cannot be performed without Sync Master being enabled.

NVAPI_SYNC_NOT_ACTIVE
The requested action cannot be performed without Sync being enabled.

NVAPI_SYSWOW64_NOT_SUPPORTED
The function is not supported from 32-bit on a 64-bit system.

NVAPI_TESTING_CLOCKS_NOT_SUPPORTED
Testing of clocks is not supported.

NVAPI_TIMEOUT
Requested operation timed out

NVAPI_TIMEOUT_RECONFIGURING_GPU_TOPO
Timeout while reconfiguring GPUs

NVAPI_TIMING_NOT_SUPPORTED
Provided timing is not supported.

NVAPI_TOO_MANY_SETTINGS_IN_PROFILE
There are too many settings for a profile.

NVAPI_TOO_MANY_UNIQUE_STATE_OBJECTS
There are too many unique instances of a particular type of state object.

NVAPI_TOPO_NOT_POSSIBLE
The mosaic topology is not possible given the current state of the hardware.

NVAPI_UNKNOWN_UNDERSCAN_CONFIG
The specified underscan config is from an unknown source (e.g. INF)

NVAPI_UNREGISTERED_RESOURCE
An unregistered resource was passed as a parameter.

NVAPI_UNSUPPORTED_CONFIG_NON_HDCP_HMD
The Config is having Non-NVidia GPU with Non-HDCP HMD connected

NVAPI_VALUE_ALREADY_SET
Value already set, setting again not allowed.

NVAPI_WAS_STILL_DRAWING
The previous operation that is transferring information to or from this surface is incomplete.

Type Definitions

NvAPI_Status
NvAPI Status Values