Enum nvapi_hi::sys::status::Status

#[repr(i32)]

pub enum Status {
    Ok,
    Error,
    LibraryNotFound,
    NoImplementation,
    ApiNotInitialized,
    InvalidArgument,
    NvidiaDeviceNotFound,
    EndEnumeration,
    InvalidHandle,
    IncompatibleStructVersion,
    HandleInvalidated,
    OpenglContextNotCurrent,
    InvalidPointer,
    NoGlExpert,
    InstrumentationDisabled,
    NoGlNsight,
    ExpectedLogicalGpuHandle,
    ExpectedPhysicalGpuHandle,
    ExpectedDisplayHandle,
    InvalidCombination,
    NotSupported,
    PortidNotFound,
    ExpectedUnattachedDisplayHandle,
    InvalidPerfLevel,
    DeviceBusy,
    NvPersistFileNotFound,
    PersistDataNotFound,
    ExpectedTvDisplay,
    ExpectedTvDisplayOnDconnector,
    NoActiveSliTopology,
    SliRenderingModeNotallowed,
    ExpectedDigitalFlatPanel,
    ArgumentExceedMaxSize,
    DeviceSwitchingNotAllowed,
    TestingClocksNotSupported,
    UnknownUnderscanConfig,
    TimeoutReconfiguringGpuTopo,
    DataNotFound,
    ExpectedAnalogDisplay,
    NoVidlink,
    RequiresReboot,
    InvalidHybridMode,
    MixedTargetTypes,
    Syswow64NotSupported,
    ImplicitSetGpuTopologyChangeNotAllowed,
    RequestUserToCloseNonMigratableApps,
    OutOfMemory,
    WasStillDrawing,
    FileNotFound,
    TooManyUniqueStateObjects,
    InvalidCall,
    D3d101LibraryNotFound,
    FunctionNotFound,
    InvalidUserPrivilege,
    ExpectedNonPrimaryDisplayHandle,
    ExpectedComputeGpuHandle,
    StereoNotInitialized,
    StereoRegistryAccessFailed,
    StereoRegistryProfileTypeNotSupported,
    StereoRegistryValueNotSupported,
    StereoNotEnabled,
    StereoNotTurnedOn,
    StereoInvalidDeviceInterface,
    StereoParameterOutOfRange,
    StereoFrustumAdjustModeNotSupported,
    TopoNotPossible,
    ModeChangeFailed,
    D3d11LibraryNotFound,
    InvalidAddress,
    StringTooSmall,
    MatchingDeviceNotFound,
    DriverRunning,
    DriverNotrunning,
    ErrorDriverReloadRequired,
    SetNotAllowed,
    AdvancedDisplayTopologyRequired,
    SettingNotFound,
    SettingSizeTooLarge,
    TooManySettingsInProfile,
    ProfileNotFound,
    ProfileNameInUse,
    ProfileNameEmpty,
    ExecutableNotFound,
    ExecutableAlreadyInUse,
    DatatypeMismatch,
    ProfileRemoved,
    UnregisteredResource,
    IdOutOfRange,
    DisplayconfigValidationFailed,
    DpmstChanged,
    InsufficientBuffer,
    AccessDenied,
    MosaicNotActive,
    ShareResourceRelocated,
    RequestUserToDisableDwm,
    D3dDeviceLost,
    InvalidConfiguration,
    StereoHandshakeNotDone,
    ExecutablePathIsAmbiguous,
    DefaultStereoProfileIsNotDefined,
    DefaultStereoProfileDoesNotExist,
    ClusterAlreadyExists,
    DpmstDisplayIdExpected,
    InvalidDisplayId,
    StreamIsOutOfSync,
    IncompatibleAudioDriver,
    ValueAlreadySet,
    Timeout,
    GpuWorkstationFeatureIncomplete,
    StereoInitActivationNotDone,
    SyncNotActive,
    SyncMasterNotFound,
    InvalidSyncTopology,
    EcidSignAlgoUnsupported,
    EcidKeyVerificationFailed,
    FirmwareOutOfDate,
    FirmwareRevisionNotSupported,
    LicenseCallerAuthenticationFailed,
    D3dDeviceNotRegistered,
    ResourceNotAcquired,
    TimingNotSupported,
    HdcpEncryptionFailed,
    PclkLimitationFailed,
    NoConnectorFound,
    HdcpDisabled,
    ApiInUse,
    NvidiaDisplayNotFound,
    PrivSecViolation,
    IncorrectVendor,
    DisplayInUse,
    UnsupportedConfigNonHdcpHmd,
    MaxDisplayLimitReached,
    InvalidDirectModeDisplay,
    GpuInDebugMode,
}

NvAPI Status Values

All NvAPI functions return one of these codes.

Variants
Ok

Success. Request is completed.

Error

Generic error

LibraryNotFound

NVAPI support library cannot be loaded.

NoImplementation

not implemented in current driver installation

ApiNotInitialized

NvAPI_Initialize has not been called (successfully)

InvalidArgument

The argument/parameter value is not valid or NULL.

NvidiaDeviceNotFound

No NVIDIA display driver, or NVIDIA GPU driving a display, was found.

EndEnumeration

No more items to enumerate

InvalidHandle

Invalid handle

IncompatibleStructVersion

An argument's structure version is not supported

HandleInvalidated

The handle is no longer valid (likely due to GPU or display re-configuration)

OpenglContextNotCurrent

No NVIDIA OpenGL context is current (but needs to be)

InvalidPointer

An invalid pointer, usually NULL, was passed as a parameter

NoGlExpert

OpenGL Expert is not supported by the current drivers

InstrumentationDisabled

OpenGL Expert is supported, but driver instrumentation is currently disabled

NoGlNsight

OpenGL does not support Nsight

ExpectedLogicalGpuHandle

Expected a logical GPU handle for one or more parameters

ExpectedPhysicalGpuHandle

Expected a physical GPU handle for one or more parameters

ExpectedDisplayHandle

Expected an NV display handle for one or more parameters

InvalidCombination

The combination of parameters is not valid.

NotSupported

Requested feature is not supported in the selected GPU

PortidNotFound

No port ID was found for the I2C transaction

ExpectedUnattachedDisplayHandle

Expected an unattached display handle as one of the input parameters.

InvalidPerfLevel

Invalid perf level

DeviceBusy

Device is busy; request not fulfilled

NvPersistFileNotFound

NV persist file is not found

PersistDataNotFound

NV persist data is not found

ExpectedTvDisplay

Expected a TV output display

ExpectedTvDisplayOnDconnector

Expected a TV output on the D Connector - HDTV_EIAJ4120.

NoActiveSliTopology

SLI is not active on this device.

SliRenderingModeNotallowed

Setup of SLI rendering mode is not possible right now.

ExpectedDigitalFlatPanel

Expected a digital flat panel.

ArgumentExceedMaxSize

Argument exceeds the expected size.

DeviceSwitchingNotAllowed

Inhibit is ON due to one of the flags in NV_GPU_DISPLAY_CHANGE_INHIBIT or SLI active.

TestingClocksNotSupported

Testing of clocks is not supported.

UnknownUnderscanConfig

The specified underscan config is from an unknown source (e.g. INF)

TimeoutReconfiguringGpuTopo

Timeout while reconfiguring GPUs

DataNotFound

Requested data was not found

ExpectedAnalogDisplay

Expected an analog display

NoVidlink

No SLI video bridge is present

RequiresReboot

NVAPI requires a reboot for the settings to take effect

InvalidHybridMode

The function is not supported with the current Hybrid mode.

MixedTargetTypes

The target types are not all the same

Syswow64NotSupported

The function is not supported from 32-bit on a 64-bit system.

ImplicitSetGpuTopologyChangeNotAllowed

There is no implicit GPU topology active. Use NVAPI_SetHybridMode to change topology.

RequestUserToCloseNonMigratableApps

Prompt the user to close all non-migratable applications.

OutOfMemory

Could not allocate sufficient memory to complete the call.

WasStillDrawing

The previous operation that is transferring information to or from this surface is incomplete.

FileNotFound

The file was not found.

TooManyUniqueStateObjects

There are too many unique instances of a particular type of state object.

InvalidCall

The method call is invalid. For example, a method's parameter may not be a valid pointer.

D3d101LibraryNotFound

d3d10_1.dll cannot be loaded.

FunctionNotFound

Couldn't find the function in the loaded DLL.

InvalidUserPrivilege

Current User is not Admin.

ExpectedNonPrimaryDisplayHandle

The handle corresponds to GDIPrimary.

ExpectedComputeGpuHandle

Setting Physx GPU requires that the GPU is compute-capable.

StereoNotInitialized

The Stereo part of NVAPI failed to initialize completely. Check if the stereo driver is installed.

StereoRegistryAccessFailed

Access to stereo-related registry keys or values has failed.

StereoRegistryProfileTypeNotSupported

The given registry profile type is not supported.

StereoRegistryValueNotSupported

The given registry value is not supported.

StereoNotEnabled

Stereo is not enabled and the function needed it to execute completely.

StereoNotTurnedOn

Stereo is not turned on and the function needed it to execute completely.

StereoInvalidDeviceInterface

Invalid device interface.

StereoParameterOutOfRange

Separation percentage or JPEG image capture quality is out of [0-100] range.

StereoFrustumAdjustModeNotSupported

The given frustum adjust mode is not supported.

TopoNotPossible

The mosaic topology is not possible given the current state of the hardware.

ModeChangeFailed

An attempt to do a display resolution mode change has failed.

D3d11LibraryNotFound

d3d11.dll/d3d11_beta.dll cannot be loaded.

InvalidAddress

Address is outside of valid range.

StringTooSmall

The pre-allocated string is too small to hold the result.

MatchingDeviceNotFound

The input does not match any of the available devices.

DriverRunning

Driver is running.

DriverNotrunning

Driver is not running.

ErrorDriverReloadRequired

A driver reload is required to apply these settings.

SetNotAllowed

Intended setting is not allowed.

AdvancedDisplayTopologyRequired

Information can't be returned due to "advanced display topology".

SettingNotFound

Setting is not found.

SettingSizeTooLarge

Setting size is too large.

TooManySettingsInProfile

There are too many settings for a profile.

ProfileNotFound

Profile is not found.

ProfileNameInUse

Profile name is duplicated.

ProfileNameEmpty

Profile name is empty.

ExecutableNotFound

Application not found in the Profile.

ExecutableAlreadyInUse

Application already exists in the other profile.

DatatypeMismatch

Data Type mismatch

ProfileRemoved

The profile passed as parameter has been removed and is no longer valid.

UnregisteredResource

An unregistered resource was passed as a parameter.

IdOutOfRange

The DisplayId corresponds to a display which is not within the normal outputId range.

DisplayconfigValidationFailed

Display topology is not valid so the driver cannot do a mode set on this configuration.

DpmstChanged

Display Port Multi-Stream topology has been changed.

InsufficientBuffer

Input buffer is insufficient to hold the contents.

AccessDenied

No access to the caller.

MosaicNotActive

The requested action cannot be performed without Mosaic being enabled.

ShareResourceRelocated

The surface is relocated away from video memory.

RequestUserToDisableDwm

The user should disable DWM before calling NvAPI.

D3dDeviceLost

D3D device status is D3DERR_DEVICELOST or D3DERR_DEVICENOTRESET - the user has to reset the device.

InvalidConfiguration

The requested action cannot be performed in the current state.

StereoHandshakeNotDone

Call failed as stereo handshake not completed.

ExecutablePathIsAmbiguous

The path provided was too short to determine the correct NVDRS_APPLICATION

DefaultStereoProfileIsNotDefined

Default stereo profile is not currently defined

DefaultStereoProfileDoesNotExist

Default stereo profile does not exist

ClusterAlreadyExists

A cluster is already defined with the given configuration.

DpmstDisplayIdExpected

The input display id is not that of a multi stream enabled connector or a display device in a multi stream topology

InvalidDisplayId

The input display id is not valid or the monitor associated to it does not support the current operation

StreamIsOutOfSync

While playing secure audio stream, stream goes out of sync

IncompatibleAudioDriver

Older audio driver version than required

ValueAlreadySet

Value already set, setting again not allowed.

Timeout

Requested operation timed out

GpuWorkstationFeatureIncomplete

The requested workstation feature set has incomplete driver internal allocation resources

StereoInitActivationNotDone

Call failed because InitActivation was not called.

SyncNotActive

The requested action cannot be performed without Sync being enabled.

SyncMasterNotFound

The requested action cannot be performed without Sync Master being enabled.

InvalidSyncTopology

Invalid displays passed in the NV_GSYNC_DISPLAY pointer.

EcidSignAlgoUnsupported

The specified signing algorithm is not supported. Either an incorrect value was entered or the current installed driver/hardware does not support the input value.

EcidKeyVerificationFailed

The encrypted public key verification has failed.

FirmwareOutOfDate

The device's firmware is out of date.

FirmwareRevisionNotSupported

The device's firmware is not supported.

LicenseCallerAuthenticationFailed

The caller is not authorized to modify the License.

D3dDeviceNotRegistered

The user tried to use a deferred context without registering the device first

ResourceNotAcquired

Head or SourceId was not reserved for the VR Display before doing the Modeset.

TimingNotSupported

Provided timing is not supported.

HdcpEncryptionFailed

HDCP Encryption Failed for the device. Would be applicable when the device is HDCP Capable.

PclkLimitationFailed

Provided mode is over sink device pclk limitation.

NoConnectorFound

No connector on GPU found.

HdcpDisabled

When a non-HDCP capable HMD is connected, we would inform user by this code.

ApiInUse

Atleast an API is still being called

NvidiaDisplayNotFound

No display found on Nvidia GPU(s).

PrivSecViolation

Priv security violation, improper access to a secured register.

IncorrectVendor

NVAPI cannot be called by this vendor

DisplayInUse

DirectMode Display is already in use

UnsupportedConfigNonHdcpHmd

The Config is having Non-NVidia GPU with Non-HDCP HMD connected

MaxDisplayLimitReached

GPU's Max Display Limit has Reached

InvalidDirectModeDisplay

DirectMode not Enabled on the Display

GpuInDebugMode

GPU is in debug mode, OC is NOT allowed.

Methods

impl Status

fn from_raw(raw: i32) -> Result<Status, ArgumentRangeError>

fn raw(&self) -> i32

fn values() -> Cloned<Iter<'static, Status>>